requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiohttp-cors==0.8.1
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1763409923949/work
attrs==26.1.0
black==24.3.0
catboost==1.2.3
category-encoders==2.6.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorful==0.5.8
colorlog==6.10.1
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
datasets==2.18.0
debugpy @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_debugpy_1769744978/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
dill==0.3.8
distlib==0.4.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
fastapi==0.135.1
filelock==3.25.2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2023.12.2
google-api-core==2.30.0
google-auth==2.49.1
googleapis-common-protos==1.73.0
graphviz==0.21
greenlet==3.3.2
grpcio==1.78.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.7.2
idna==3.11
imbalanced-learn==0.12.3
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipykernel_1770566133/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1772790149/work
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jupyter_client_1767954655/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jupyter_core_1760643864/work
kiwisolver==1.5.0
lightgbm==4.6.0
linkify-it-py==2.1.0
llvmlite==0.42.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1761214490209/work
mdit-py-plugins==0.5.0
mdurl==0.1.2
memray==1.19.2
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==2.18.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
numba==0.59.1
numpy==1.26.4
nvidia-nccl-cu12==2.29.7
opencensus==0.11.4
opencensus-context==0.1.3
optuna==3.6.2
oracle-automlx==25.3.1
packaging @ file:///home/task_176104874243446/conda-bld/packaging_1761049080023/work
pandas==2.2.2
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1770676243/work
pathspec==1.0.4
patsy==1.0.2
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pillow==10.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1773199142/work
plotly==6.6.0
prometheus_client==0.24.1
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.6
psutil==6.1.1
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
py-spy==0.4.1
pyarrow==15.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pyod==1.1.3
pyparsing==3.3.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-discovery==1.2.0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pyzmq_1771716961/work
ray==2.32.0
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
scikit-learn==1.3.2
scipy==1.13.1
seaborn==0.13.2
setuptools==69.5.1
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
smart_open==7.5.1
SQLAlchemy==2.0.48
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==1.0.0
statsmodels==0.14.6
tensorboardX==2.6.4
textual==8.1.1
threadpoolctl==3.6.0
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1765836395280/work
tqdm==4.67.3
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==2.0.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
virtualenv==21.2.0
watchfiles==1.1.1
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1770634525285/work
websockets==16.0
wheel==0.46.3
wrapt==2.1.2
xgboost==2.1.3
xxhash==3.6.0
yarl==1.23.0
